+++ Sources/Demo/AppDelegate.swift
@@ -14,6 +14,34 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
             ServiceProvider.shared.localHistoryProvider.deleteAll()
         }
 
+        setUpProgrammaticUI(application: application)
+
         return true
+    }
+
+    /// Set up the fifth tab 'Offline' programmatically.
+    private func setUpProgrammaticUI(application: UIApplication) {
+        let tabBarControllers = application.connectedScenes
+            .compactMap { $0 as? UIWindowScene }
+            .flatMap(\.windows)
+            .compactMap { window in
+                window.backgroundColor = UIColor.systemBackground
+                return window.rootViewController as? UITabBarController
+            }
+
+        let offlineDemoViewController = OfflineDemoViewController()
+        offlineDemoViewController.tabBarItem = UITabBarItem(
+            title: "Offline",
+            image: UIImage(systemName: "icloud.and.arrow.down"),
+            tag: 0
+        )
+
+        for tabBarController in tabBarControllers {
+            guard var viewControllers = tabBarController.viewControllers else {
+                break
+            }
+            viewControllers.append(offlineDemoViewController)
+            tabBarController.setViewControllers(viewControllers, animated: false)
+        }
     }
 }